修复msg id为Int32, 加入撤回消息功能

+++ SaladimQBot.Core/-Interfaces/IMessage.cs
@@ -2,7 +2,7 @@ namespace SaladimQBot.Core;
 
 public interface IMessage : IClientEntity
 {
-    long MessageId { get; }
+    int MessageId { get; }
 
     IMessageEntity MessageEntity { get; }
